# Replace matrix-matrix mul with matrix-vector mul in `dssesolve`