Previous implementation of `dssesolve` was doing an unnecessary matrix-matrix multiplication. 
It can be replaced with two matrix-vector multiplication, which is more efficient. 
A small optimization can be carried out for the computation of the `exp` but I don't understand how all those trailing dimension should be handled.