@pierreguilmin 2025-02-20
That's great, thanks a lot. 😍 I'll let you review my commit, and merge if it's correct.

Speedup
| n | ntrajs      | Time for `main`   | New time  | Speedup     |
|------|--------------|-------------|------------|-------------|
| 32 x 8 | 100        | 1.25 s     | 270 ms     | **x4.6**        |
| 32 x 8 | 1000       | 11.5 s     | 1.62 s     | **x7.2**          |

<details>
  <summary>Benchmark code</summary>

```python
import jax
import jax.numpy as jnp
import dynamiqs as dq

na = 32
nb = 8
a, b = dq.destroy(na, nb)

g2 = 1.0
eb = -3.0
kb = 10.0
H = g2 * a @ a @ b.dag() + dq.hc + (eb * b + dq.hc)
jump_ops = [jnp.sqrt(kb) * b]
psi0 = dq.coherent((na, nb), (0.0, 0.0))
tsave = jnp.linspace(0, 2.0, 101)

key = jax.random.PRNGKey(42)
ntrajs = 1000
keys = jax.random.split(key, ntrajs)
method = dq.method.EulerMaruyama(dt=1e-3)
psi_sse = dq.dssesolve(H, jump_ops, psi0, tsave, keys, method=method).states

%timeit -r1 -n1 -q dq.dssesolve(H, jump_ops, psi0, tsave, keys, method=method).states.block_until_ready()
%timeit dq.dssesolve(H, jump_ops, psi0, tsave, keys, method=method).states.block_until_ready()
```
</details>