@gaspardbb 2025-02-21
Looks great, thanks! So it seems this was an optimization the jax compiler missed 💪